Add files via upload
This project measures electricity consumption and bills according to eThekwini municipality tariff for 2024. The project measures the consumption across one load but can be expanded to more than one load. This project can also be modified to be a maximum demand controller, and with IoT it can be a smart electricity meter.